requirements.txt
groq
openai-whisper
python-dotenv